pyproject.toml
[project]
name = "Guido"
version = "0.1.0"
description = "Everyone's favourite Cars character, now as an autonomous vehicle planner."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "pandas>=3.0.2",
    "pyyaml>=6.0.3",
    "termcolor>=3.3.0",
    "torch",
    "torchmetrics>=1.9.0",
    "torchvision",
    "transformers>=4.56.0",
]


[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu126" }
torchvision = { index = "pytorch-cu126" }
